roles

+++ app/Http/Controllers/RoomController.php
@@ -9,6 +9,7 @@ use App\Models\Image;
 use App\Models\Room;
 use App\Models\RoomType;
 use Exception;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Storage;
 
@@ -27,50 +28,55 @@ class RoomController extends Controller
     public function store(RoomRequest $request)
     {
         try {
-            DB::beginTransaction();
+            if (auth::check()) {
+                DB::beginTransaction();
 
-            $room = Room::create(array_merge(
-                $request->validated(),
-                ['user_id' => auth()->id()]
-            ));
+                $room = Room::create(array_merge(
+                    $request->validated(),
+                    ['user_id' => auth()->id()]
+                ));
 
-            if ($request->has('features')) {
-                foreach ($request->input('features') as $featureName) {
-                    $feature = Feature::firstOrCreate([
-                        'name' => $featureName,
-                    ]);
+                if ($request->has('features')) {
+                    foreach ($request->input('features') as $featureName) {
+                        $feature = Feature::firstOrCreate([
+                            'name' => $featureName,
+                        ]);
 
-                    $room->features()->attach($feature->id);
+                        $room->features()->attach($feature->id);
+                    }
                 }
-            }
 
-            if ($request->hasFile('image')) {
-                $path = $request->file('image')->store('rooms', 'public');
+                if ($request->hasFile('image')) {
+                    $path = $request->file('image')->store('rooms', 'public');
 
-                Image::create([
-                    'room_id' => $room->id,
-                    'image_path' => $path,
-                    'name' => $request->file('image')->getClientOriginalName(),
+                    Image::create([
+                        'room_id' => $room->id,
+                        'image_path' => $path,
+                        'name' => $request->file('image')->getClientOriginalName(),
+                    ]);
+                }
+
+                if ($request->has('room_type')) {
+                    $roomType = $request->input('room_type');
+
+                    RoomType::create([
+                        'room_id' => $room->id,
+                        'name' => $roomType['name'],
+                        'description' => $roomType['description'],
+                    ]);
+                }
+
+                DB::commit();
+
+                return ApiResponse::success([
+                    'data' => $room->load(['images', 'features', 'roomType']),
+                    'message' => 'Room created successfully.',
+                ]);
+            } else {
+                return ApiResponse::error([
+                    'message' => 'Unauthorized'
                 ]);
             }
-
-            if ($request->has('room_type')) {
-                $roomType = $request->input('room_type');
-
-                RoomType::create([
-                    'room_id' => $room->id,
-                    'name' => $roomType['name'],
-                    'description' => $roomType['description'],
-                ]);
-            }
-
-            DB::commit();
-
-            return ApiResponse::success([
-                'data' => $room->load(['images', 'features', 'roomType']),
-                'message' => 'Room created successfully.',
-            ]);
-
         } catch (Exception $e) {
             DB::rollBack();
 
@@ -79,6 +85,7 @@ class RoomController extends Controller
             ]);
         }
     }
+
 
     public function show(string $id)
     {
@@ -99,86 +106,92 @@ class RoomController extends Controller
     public function update(RoomRequest $request, string $id)
     {
         try {
-            DB::beginTransaction();
+            if (auth::check()) {
+                DB::beginTransaction();
 
-            $room = Room::find($id);
+                $room = Room::find($id);
 
-            if($room->user_id != auth()->id()){
-                return ApiResponse::error([
-                    'message' => 'Unauthorized.'
-                ]);
-            }
+                if($room->user_id != auth()->id()){
+                    return ApiResponse::error([
+                        'message' => 'Unauthorized.'
+                    ]);
+                }
 
-            if (!$room) {
-                return ApiResponse::error([
-                    'message' => 'Room not found.'
-                ]);
-            }
+                if (!$room) {
+                    return ApiResponse::error([
+                        'message' => 'Room not found.'
+                    ]);
+                }
 
-            $room->update($request->validated());
+                $room->update($request->validated());
 
-            if ($request->hasFile('image') && $request->file('image')->isValid()) {
-                $path = $request->file('image')->store('rooms', 'public');
+                if ($request->hasFile('image') && $request->file('image')->isValid()) {
+                    $path = $request->file('image')->store('rooms', 'public');
 
-                $image = Image::where('room_id', $room->id)->first();
+                    $image = Image::where('room_id', $room->id)->first();
 
-                if ($image && $image->image_path) {
-                    if (Storage::disk('public')->exists($image->image_path)) {
-                        Storage::disk('public')->delete($image->image_path);
+                    if ($image && $image->image_path) {
+                        if (Storage::disk('public')->exists($image->image_path)) {
+                            Storage::disk('public')->delete($image->image_path);
+                        }
                     }
-                }
 
-                if ($image) {
-                    $image->update([
-                        'image_path' => $path,
-                        'name' => $request->file('image')->getClientOriginalName()
-                    ]);
-                } else {
-                    Image::create([
-                        'room_id' => $room->id,
-                        'image_path' => $path,
-                        'name' => $request->file('image')->getClientOriginalName()
-                    ]);
-                }
-            }
-
-            if ($request->has('room_type')) {
-                $details = $request->input('room_type');
-
-                if ($details) {
-                    $roomType = RoomType::where('room_id', $room->id)->first();
-
-                    if ($roomType) {
-                        $roomType->update([
-                            'name' => $details['name'],
-                            'description' => $details['description'],
+                    if ($image) {
+                        $image->update([
+                            'image_path' => $path,
+                            'name' => $request->file('image')->getClientOriginalName()
                         ]);
                     } else {
-                        RoomType::create([
+                        Image::create([
                             'room_id' => $room->id,
-                            'name' => $details['name'],
-                            'description' => $details['description'],
+                            'image_path' => $path,
+                            'name' => $request->file('image')->getClientOriginalName()
                         ]);
                     }
                 }
-            }
 
-            if ($request->has('features')) {
-                $featureIds = [];
-                foreach ($request->input('features') as $featureName) {
-                    $feature = Feature::firstOrCreate(['name' => $featureName]);
-                    $featureIds[] = $feature->id;
+                if ($request->has('room_type')) {
+                    $details = $request->input('room_type');
+
+                    if ($details) {
+                        $roomType = RoomType::where('room_id', $room->id)->first();
+
+                        if ($roomType) {
+                            $roomType->update([
+                                'name' => $details['name'],
+                                'description' => $details['description'],
+                            ]);
+                        } else {
+                            RoomType::create([
+                                'room_id' => $room->id,
+                                'name' => $details['name'],
+                                'description' => $details['description'],
+                            ]);
+                        }
+                    }
                 }
 
-                $room->features()->sync($featureIds);
+                if ($request->has('features')) {
+                    $featureIds = [];
+                    foreach ($request->input('features') as $featureName) {
+                        $feature = Feature::firstOrCreate(['name' => $featureName]);
+                        $featureIds[] = $feature->id;
+                    }
+
+                    $room->features()->sync($featureIds);
+                }
+
+                DB::commit();
+
+                return ApiResponse::success([
+                    'data' => $room->load(['images', 'features', 'roomType']),
+                    'message' => 'Room updated successfully.'
+                ]);
+            }else {
+                return ApiResponse::error([
+                    'message' => 'Unauthorized'
+                ]);
             }
-
-            DB::commit();
-
-            return ApiResponse::success([
-                'data' => $room->load(['images', 'features', 'roomType']),
-                'message' => 'Room updated successfully.'
-            ]);
         } catch (Exception $e) {
             DB::rollBack();
 
@@ -190,24 +203,30 @@ class RoomController extends Controller
 
     public function destroy(string $id)
     {
-        $room = Room::find($id);
+        if(auth::check()){
+            $room = Room::find($id);
 
-        if ($room->user_id != auth()->id()) {
-           return ApiResponse::error([
-               'message' => 'Unauthorized.'
-           ]);
-        }
+            if ($room->user_id != auth()->id()) {
+                return ApiResponse::error([
+                    'message' => 'Unauthorized.'
+                ]);
+            }
 
-        if (!$room) {
+            if (!$room) {
+                return ApiResponse::error([
+                    'message' => 'Room not found.'
+                ]);
+            }
+            $room->delete();
+
+            return ApiResponse::success([
+                'data' => $room,
+                'message' => 'Room deleted successfully.'
+            ]);
+        }else{
             return ApiResponse::error([
-                'message' => 'Room not found.'
+                'message' => 'Unauthorized'
             ]);
         }
-        $room->delete();
-
-        return ApiResponse::success([
-            'data' => $room,
-            'message' => 'Room deleted successfully.'
-        ]);
     }
 }
